Rename containers to pages in email
