+++ packages/core/email/admin/src/pages/Settings/components.js
