+++ packages/core/email/admin/src/pages/Settings/index.js
